+++ docsy.dev/content/en/project/about/changelog.md
@@ -133,6 +133,8 @@ For the full list of changes, see the [0.15.1][] or [0.16.0][] release page.
 
 [**Breaking changes**](#breaking-change):
 
+- **Default favicons** removed from the theme; sites now supply their own. See
+  [Add your favicons][favicons] ([#2653][]).
 - Raised the theme's minimum supported Hugo version to
   **[0.158.0][hugo-0.158.0]** (was 0.146.0; Docsy 0.15 documented Hugo 0.157.0).
   Theme templates now use the language APIs introduced in Hugo 0.158.0. Note:
@@ -149,7 +151,9 @@ For the full list of changes, see the [0.15.1][] or [0.16.0][] release page.
   minimum supported Hugo version remains 0.158.0.
 
 [#2594]: https://github.com/google/docsy/pull/2594
+[#2653]: https://github.com/google/docsy/pull/2653
 [#2578]: https://github.com/google/docsy/pull/2578
+[favicons]: /docs/content/iconsimages/#add-your-favicons
 [hugo-0.163.1]: https://github.com/gohugoio/hugo/releases/tag/v0.163.1
 
 [**Experimental**](#experimental):